safe undelete

+++ lib/app/helpers/inline_forms_helper.rb
@@ -32,14 +32,34 @@ module InlineFormsHelper
 
   # destroy link
   def link_to_destroy( object, update_span )
-    if cancan_disabled? || ( can? :delete, object )
-      link_to "<i class='fi-trash'></i>".html_safe,
-        send( object.class.to_s.underscore + '_path',
-        object,
-        :update => update_span ),
-        :method => :delete,
-        :remote => true,
-        :title => t('inline_forms.view.trash')
+    if current_user.role? :superadmin
+      if cancan_disabled? || ( can? :delete, object )
+        link_to "<i class='fi-trash'></i>".html_safe,
+          send( object.class.to_s.underscore + '_path',
+          object,
+          :update => update_span ),
+          :method => :delete,
+          :remote => true,
+          :title => t('inline_forms.view.trash')
+      end
+    elsif (object.class.safe_deletable? rescue false)
+      if object.deleted?     && (cancan_disabled? || ( can? :revert, object ))
+        link_to "undelete".html_safe,
+          send( 'revert_' + object.class.to_s.underscore + '_path',
+          object,
+          :update => update_span ),
+          :method => :post,
+          :remote => true,
+          :title => t('inline_forms.view.trash')
+      elsif !object.deleted? && (cancan_disabled? || ( can? :delete, object ))
+        link_to "<i class='fi-trash'></i>".html_safe,
+          send( object.class.to_s.underscore + '_path',
+          object,
+          :update => update_span ),
+          :method => :delete,
+          :remote => true,
+          :title => t('inline_forms.view.trash')
+      end
     end
   end
